refactor(parser): move parser functionalities to js
This PR deliverately introduces a breaking API change.
