+++ js/builder.go
@@ -1,7 +1,6 @@
 package js
 
 import (
-	"github.com/xjslang/xjs/lexer"
 	"github.com/xjslang/xjs/parser"
 )
 
@@ -9,8 +8,7 @@ type Builder struct {
 	parser.Builder
 }
 
-// TODO: https://github.com/xjslang/xjs/pull/90#discussion_r3191186051
-func (b *Builder) Build(l *lexer.Lexer) *parser.Parser {
+func (b *Builder) InstallCorePlugins() {
 	b.Install(LetPlugin)
-	return b.Builder.Build(l)
+	b.Install(FunctionPlugin)
 }
