+++ parser/scope_tracker.go
@@ -3,20 +3,20 @@ package parser
 type scope int
 
 const (
-	blockScope scope = iota
+	BlockScope scope = iota
 )
 
 type scopeTracker map[scope]int
 
-func (st scopeTracker) enter(sc scope) {
+func (st scopeTracker) Enter(sc scope) {
 	st[sc]++
 }
 
-func (st scopeTracker) in(sc scope) bool {
+func (st scopeTracker) In(sc scope) bool {
 	return st[sc] > 0
 }
 
-func (st scopeTracker) exit(sc scope) {
+func (st scopeTracker) Exit(sc scope) {
 	st[sc]--
 	if st[sc] <= 0 {
 		delete(st, sc)
